% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My Bin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{number of iterations}

\item{n}{trials}

\item{p}{probability}
}
\value{
a binomial simulation with x number of successes
}
\description{
My Bin
}
\examples{
mybin(iter=100,n=10, p=0.7)
}
