% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{My Boot 2}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{number of iterations}

\item{x}{sample}

\item{fun}{title}

\item{alpha}{alpha}

\item{cx}{cx}

\item{...}{}
}
\value{
BootStrap
}
\description{
My Boot 2
}
