% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{My Confidence Interval}
\usage{
myci(x)
}
\arguments{
\item{x}{single sample x}
}
\value{
a 95% confidence interval for the mean from a single sample x
}
\description{
My Confidence Interval
}
\examples{
x = rnorm(30, mean=10, sd=12); myci(x)
}
