% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{My N Curve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{mean}

\item{sigma}{standard deviation}

\item{a}{y value}
}
\value{
the curve, the shaded area between the curve and x-axis from -infinity to x=a, and the area
}
\description{
My N Curve
}
\examples{
myncurve(mu = 0, sigma = 1, a = 2)
}
