% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mypvalue.R
\name{mypvalue}
\alias{mypvalue}
\title{My P Value}
\usage{
mypvalue(t0, xmax = 4, n = 20, alpha = 0.05)
}
\arguments{
\item{t0}{the calculated t-value from a statistical test}

\item{xmax}{the max x-value on the plot for the t distribution}

\item{n}{sample size}

\item{alpha}{significance level for the test}
}
\value{
a list containing q and the p-value, as well as a graph showing the rejection areas of the distribution
}
\description{
My P Value
}
\examples{
x1=rnorm(30,mean=25,sd=5)
tcalc = (mean(x1)-24)/(sd(x1)/sqrt(30))
mypvalue(tcalc, n=30, alpha=0.05)
}
