% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nTickets.R
\name{nTickets}
\alias{nTickets}
\title{N Tickets}
\usage{
nTickets(N, gamma, p)
}
\arguments{
\item{N}{number of seats in flight}

\item{gamma}{probability of overbooking}

\item{p}{probability a passenger will show}
}
\value{
prints a named list containing nd, nc, N, p and gamma - where nd is calculated using the discrete distribution and nc is the same calculated with normal approximation. Also creates discrete and continuous plots of Objective function vs n
}
\description{
N Tickets
}
\examples{
nTickets(400, 0.02, 0.95)
}
