% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pareto.R
\name{pareto}
\alias{pareto}
\title{Pareto Barplot}
\usage{
pareto(x, mn = "Pareto barplot")
}
\arguments{
\item{x}{a vector of relative frequency}

\item{mn}{title}
}
\value{
a pareto barplot
}
\description{
Pareto Barplot
}
\examples{
freq <- c(15, 8, 63, 20)
RL = c("None", "Both", "LegsOnly", "WheelsOnly")
l = rep(RL, freq)
pareto(l)
}
